requirements.txt
torch>=2.0.0
transformers>=4.30.0
pillow>=9.0.0
pandas>=1.5.0
numpy>=1.21.0
tqdm>=4.65.0 